changed to gradle
