+++ build.gradle
@@ -1,0 +1,35 @@
+plugins {
+	id 'org.springframework.boot' version '2.3.2.RELEASE'
+	id 'io.spring.dependency-management' version '1.0.9.RELEASE'
+	id 'java'
+	id 'war'
+}
+
+group = 'cz.vse.fcp'
+version = '0.0.1-SNAPSHOT'
+sourceCompatibility = '1.8'
+
+repositories {
+	mavenCentral()
+}
+
+dependencies {
+	compile group: 'net.sourceforge.owlapi', name: 'jfact', version: '5.0.3'
+	compile group: 'net.sourceforge.owlapi', name: 'owlapi-distribution', version: '5.1.16'
+//	compile group: 'com.fasterxml.jackson.core', name: 'jackson-annotations', version: '2.8.0'
+//	compile group: 'com.fasterxml.jackson.core', name: 'jackson-core', version: '2.8.9'
+//	compile group: 'com.fasterxml.jackson.core', name: 'jackson-databind', version: '2.8.9'
+//	compile group: 'com.fasterxml.jackson.dataformat', name: 'jackson-dataformat-xml'
+//	compile group: 'com.fasterxml.jackson.dataformat', name: 'jackson-dataformat-cbor'
+//	compile group: 'com.fasterxml.jackson.dataformat', name: 'jackson-dataformat-smile'
+
+	implementation 'org.springframework.boot:spring-boot-starter-web'
+	providedRuntime 'org.springframework.boot:spring-boot-starter-tomcat'
+	testImplementation('org.springframework.boot:spring-boot-starter-test') {
+		exclude group: 'org.junit.vintage', module: 'junit-vintage-engine'
+	}
+}
+
+test {
+	useJUnitPlatform()
+}
