+++ src/main/java/cz/vse/fcp/demo/DemoApplication.java
@@ -1,0 +1,13 @@
+package cz.vse.fcp.demo;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class DemoApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(DemoApplication.class, args);
+	}
+
+}
