+++ src/test/java/cz/vse/fcp/demo/DemoApplicationTests.java
@@ -1,0 +1,13 @@
+package cz.vse.fcp.demo;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class DemoApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
